Initial commit for schotime/NPoco#150
1. Created OutoutColumnAttribute
2. Created Unit Test for Insert with OutputColumns
3. Extended unit test database schema with additional server generated columns for testing this stuff
4. Created Unit Test POCO for testing this attribute
5. ColumnInfo now populated with OutputColumnNames for INSERT and UPDATE
6. PocoColumn and PocoData now have information about OutputColumn names for INSERT and UPDATE
7. Modified Database.cs INSERT method, trying to get it to handle OUTPUT columns

Still to do:

MySql doesn't support OUTPUT columns, and other ones may not either.

Firebird provider already uses a "Returning" clause to get the inserted id, and adding output paramaters is just a matter of including additional column names in this clause. This is different from the T-SQL based providers that use an OUTPUT clause, but also use a seperate SELECT with Scope_Identity or something to get the ID as a seperate result.

Therefore it's pretty clear that depending upon the provider:

1. Output columns may not be supported at all.
2. The Output values may be returned in a DataReader, or as Return paramaters on the command
3. If using AutoIncrementalId, the ID value may be returned along with the Output params in the datareader, or as a Return parameter on the command, or as a seperate Result set (e.g where a seperate SELECT is used)

Due to the different ways the id, and the output paramaetrs can be handled, it's hard trying to think of the nicest contract between the DatabaseType's and the Database.cs class. At the moment I am trying to keep it as less granular as possible.. Open to suggestions.